Refactor soft navigation script into modules
