+++ Pages/Shared/_Layout.cshtml
@@ -94,6 +94,6 @@
     <div id="pageScripts">
         @RenderSection("Scripts", required: false)
     </div>
-    <script src="~/js/soft-navigation.js" asp-append-version="true"></script>
+    <script type="module" src="~/js/soft-navigation/index.js" asp-append-version="true"></script>
 </body>
 </html>
